test case for Issue #30
prebuild creates a small JAR library file outside of maven then uses
`mvn install:install-file` to put it into the IT localRepositoryPath
without a POM.

pom.xml has the JAR library as a dependency... the IT build phase
emits WARNING about the missing POM (see build.log), but completes
as this is not a maven error.

verify calls `mvn project-info-reports:dependencies` and this is
where the NullPointerException is thrown (see verify.log - I followed
the instructions there to enable full debug logging and that's what
led me to JGitverModelProcessor.provisionModel).