make javadoc work with java8
currently ther  are two problems. first is with diffplus spotless,
which fails the formatting when using java8:
https://github.com/diffplug/spotless/issues/658

second, javadoc fails when using java9 and newer when not setting
JAVA_HOME like on many linux distros, and macos:
https://github.com/apache/maven-javadoc-plugin/commit/4bf64b02

this will be only fixed with the maven-javadoc-plugin- 3.2.1
whenever it will be released.